add bias, mix dma sound

+++ eggvance/src/apu/fifo.h
@@ -4,7 +4,7 @@
 
 #include "base/int.h"
 
-class Fifo : public shell::RingBuffer<u8, 32>
+class Fifo : public shell::RingBuffer<s8, 32>
 {
 public:
     bool refillable() const
@@ -12,5 +12,5 @@ public:
         return size() <= 16;
     }
 
-    s16 sample = 0;
+    s8 sample = 0;
 };
